// ClientWeb/src/app/model/jucator.ts
import { IEchipa } from "./echipa";
import { IPremiiP } from "./premii-persoane";

export interface IJucator{
    id: number;
    nume: string;
    prenume: string;
    dataNasterii: string;
    inaltime: number;
    post: string;
    descriere: string;
    nationalitate: string;
    linkPoza: string;
    premii: IPremiiP[];
    roluri: Map<Date, string>;
    lotSeniori: number;
    echipa: IEchipa;
}